refactor: simplify reset password form and add render method
The Livewire ResetPassword component was refactored to simplify the form and improve its structure.

- Removed unnecessary extra attributes from form fields, relying on Filament's internal handling for names and defaults. This cleans up the code and reduces redundancy.

- Added a render method to explicitly define the view to be rendered. This makes the component more self-contained and improves readability.

- Moved password reset related tests from `tests/Feature/PasswordResetTest.php` to `tests/Feature/Livewire/Auth/ResetPasswordTest.php` and refactored them to use the Livewire test helper.
