+++ app/Livewire/Auth/ResetPassword.php
@@ -10,6 +10,7 @@ use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Forms\Contracts\HasForms;
 use Filament\Forms\Form;
 use Filament\Support\Enums\Alignment;
+use Illuminate\Contracts\View\View;
 use Illuminate\Validation\Rules\Password;
 use Livewire\Component;
 
@@ -40,15 +41,12 @@ class ResetPassword extends Component implements HasForms
                 Section::make('request-password')
                     ->heading(__('filament-panels::pages/auth/password-reset/reset-password.heading'))
                     ->schema([
-                        Hidden::make('token')
-                            ->extraAttributes(['name' => 'token']),
+                        Hidden::make('token'),
                         TextInput::make('email')
                             ->label(__('filament-panels::pages/auth/password-reset/reset-password.form.email.label'))
                             ->readOnly()
                             ->autofocus()
-                            ->autocomplete('username')
-                            ->default(old('email'))
-                            ->extraInputAttributes(['name' => 'email']),
+                            ->autocomplete('username'),
                         TextInput::make('password')
                             ->label(__('filament-panels::pages/auth/password-reset/reset-password.form.password.label'))
                             ->password()
@@ -56,15 +54,13 @@ class ResetPassword extends Component implements HasForms
                             ->required()
                             ->rule(Password::default())
                             ->same('passwordConfirmation')
-                            ->validationAttribute(__('filament-panels::pages/auth/password-reset/reset-password.form.password.validation_attribute'))
-                            ->extraInputAttributes(['name' => 'password']),
+                            ->validationAttribute(__('filament-panels::pages/auth/password-reset/reset-password.form.password.validation_attribute')),
                         TextInput::make('passwordConfirmation')
                             ->label(__('filament-panels::pages/auth/password-reset/reset-password.form.password_confirmation.label'))
                             ->password()
                             ->revealable()
                             ->required()
-                            ->dehydrated(false)
-                            ->extraInputAttributes(['name' => 'password_confirmation']),
+                            ->dehydrated(false),
                     ])
                     ->footerActions([
                         Action::make('resetPassword')
@@ -74,5 +70,10 @@ class ResetPassword extends Component implements HasForms
                     ->footerActionsAlignment(Alignment::Right),
             ])
             ->statePath('data');
+    }
+
+    public function render(): View
+    {
+        return view('livewire.auth.reset-password');
     }
 }
