+++ tests/Feature/Livewire/Auth/ResetPasswordTest.php
@@ -3,14 +3,202 @@
 namespace Tests\Feature\Livewire\Auth;
 
 use App\Livewire\Auth\ResetPassword;
+use App\Models\User;
+use Illuminate\Auth\Notifications\ResetPassword as NotificationsResetPassword;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Support\Facades\Notification;
+use Laravel\Fortify\Features;
 use Livewire\Livewire;
 use Tests\TestCase;
 
 class ResetPasswordTest extends TestCase
 {
+    use RefreshDatabase;
+
     public function test_reset_password_can_be_rendered(): void
     {
         Livewire::test(ResetPassword::class)
             ->assertStatus(200);
+    }
+
+    public function test_reset_password_link_screen_can_be_rendered(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        $response = $this->get('/forgot-password');
+
+        $response->assertStatus(200);
+    }
+
+    public function test_reset_password_link_can_be_requested(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class);
+    }
+
+    public function test_reset_password_screen_can_be_rendered(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class, function (NotificationsResetPassword $notification) {
+            $response = $this->get('/reset-password/'.$notification->token);
+
+            $response->assertStatus(200);
+
+            return true;
+        });
+    }
+
+    public function test_password_can_be_reset_with_valid_token(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class, function (NotificationsResetPassword $notification) use ($user) {
+            $response = $this->post('/reset-password', [
+                'token' => $notification->token,
+                'email' => $user->email,
+                'password' => 'password',
+                'password_confirmation' => 'password',
+            ]);
+
+            $response->assertSessionHasNoErrors();
+
+            return true;
+        });
+    }
+
+    public function test_password_reset_fails_with_invalid_token(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        $user = User::factory()->create();
+
+        $response = $this->post('/reset-password', [
+            'token' => 'invalid-token',
+            'email' => $user->email,
+            'password' => 'password',
+            'password_confirmation' => 'password',
+        ]);
+
+        $response->assertSessionHasErrors(['email' => 'This password reset token is invalid.']);
+    }
+
+    public function test_password_reset_fails_with_password_mismatch(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class, function (NotificationsResetPassword $notification) use ($user) {
+            $response = $this->post('/reset-password', [
+                'token' => $notification->token,
+                'email' => $user->email,
+                'password' => 'password',
+                'password_confirmation' => 'different-password',
+            ]);
+
+            $response->assertSessionHasErrors(['password']);
+
+            return true;
+        });
+    }
+
+    public function test_password_reset_fails_with_password_too_short(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class, function (NotificationsResetPassword $notification) use ($user) {
+            $response = $this->post('/reset-password', [
+                'token' => $notification->token,
+                'email' => $user->email,
+                'password' => 'short',
+                'password_confirmation' => 'short',
+            ]);
+
+            $response->assertSessionHasErrors(['password']);
+
+            return true;
+        });
+    }
+
+    public function test_password_reset_redirects_after_successful_reset(): void
+    {
+        if (! Features::enabled(Features::resetPasswords())) {
+            $this->markTestSkipped('Password updates are not enabled.');
+        }
+
+        Notification::fake();
+
+        $user = User::factory()->create();
+
+        $this->post('/forgot-password', [
+            'email' => $user->email,
+        ]);
+
+        Notification::assertSentTo($user, NotificationsResetPassword::class, function (NotificationsResetPassword $notification) use ($user) {
+            $response = $this->post('/reset-password', [
+                'token' => $notification->token,
+                'email' => $user->email,
+                'password' => 'password',
+                'password_confirmation' => 'password',
+            ]);
+
+            $response->assertRedirect('/login');
+
+            return true;
+        });
     }
 }
